devos-mvp/requirements.txt
# DevOS MVP Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AWS Integration
boto3==1.34.0
botocore==1.34.0

# Database
sqlalchemy==2.0.23
sqlite3
alembic==1.13.1

# System Monitoring
psutil==5.9.6
watchdog==3.0.0

# File Processing
aiofiles==23.2.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Async Support
asyncio
aiohttp==3.9.1

# Utilities
pathlib
datetime
hashlib
uuid

# Optional Dependencies for Enhanced Features
# Speech Recognition (optional)
# speechrecognition==3.10.0
# pyaudio==0.2.11

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1